requirements.txt
Django>=5.1.1,<5.2
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.1,<2.10
drf-spectacular>=0.27.2,<0.28
django-allauth>=65.0.2,<65.1
django-allauth[google]>=65.0.2,<65.1
pyjwt>=2.9.0,<2.10
requests>=2.32.3,<2.33
cryptography>=43.0.1,<43.1
dj-rest-auth>=6.0.0,<6.1
drf-spectacular>=0.27.2,<0.28
django-extensions>=3.2.3,<3.3
django-cors-headers>=4.5.0,<4.6
google-genai
httpx==0.27.2
